+++ src/Modules/Search/SearchTransactions.php
@@ -69,7 +69,7 @@ class SearchTransactions
 
         if ($this->featureToggleChecker->isFeatureToggleActive(FeatureToggleDefinitions::MAIL_SEARCH->value)) {
             $start = microtime(true);
-            $boxes = $this->mailboxGateway->getBoxes($this->currentUserUnits->isAdminFor(null), $foodsaverId, $this->session->mayRole(Role::STORE_MANAGER));
+            $boxes = $this->mailboxGateway->getBoxes($this->currentUserUnits->isAdminFor(null), $foodsaverId);
             $mailboxIds = array_column($boxes, 'id');
             $result->mails = $this->searchGateway->searchMails($query, $mailboxIds);
             $result->timings['mails'] = microtime(true) - $start;
@@ -91,7 +91,7 @@ class SearchTransactions
     public function searchIndex(): MixedSearchResult
     {
         $foodsaverId = $this->session->id();
-        $boxes = $this->mailboxGateway->getBoxes($this->currentUserUnits->isAdminFor(null), $foodsaverId, $this->session->mayRole(Role::STORE_MANAGER));
+        $boxes = $this->mailboxGateway->getBoxes($this->currentUserUnits->isAdminFor(null), $foodsaverId);
         $mailboxIds = array_column($boxes, 'id');
 
         $result = new MixedSearchResult();