+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Userstorage saved to Localstoage !3621 @SaiCode
 - Ambassadors can choose if user will be informed with bell and mail when creating ID cards #2313 !3995 @McGoldi
 - Update team Page (linking / design) !3999 @SaiCode
+- Create a personal immediately after finishing the store manager quiz !3983 @alex.simm
 
 ## Bugfixes
 - Resolve "region filter bug on profile page: Blank screen when navigating store list" #2295 !3940 @chriswalg